part.py: print joystick capabilities in verbose mode
When the user starts the script with log cli option, it displays the
key presses to the console. This info is helpful in knowing the
key-mappings for a new controller.

However some details like maximum and minimum values for absolute
axes are missing. Displaying capabilities will help resolve this.

Here is a snippet from this:
...
{('EV_ABS', 3): [(('ABS_X', 0),
                  AbsInfo(value=-386, min=-32768, max=32767, fuzz=16, flat=128, resolution=0)),
...

Signed-off-by: Jithu Joseph <jithu83@gmail.com>